Alias pinned nightly to "normal" nightly
As part of the ink! CI we have some stages which use the `+nightly`
calling convention. These ended up failing since we didn't have a
nightly toolchain, only a `+nightly-zzzz-yy-xx` one.

In the past we've had both toolchains installed, but this meant possible
inconsistencies when running those stages.

This takes inspiration from the base CI Dockerfile.