ci-linux: Revert nightly rust to 2022-11-16
The current nightlies introduce a ton of new clippy warnings, and their
clippy is also mildly buggy, which makes it a pain to fix everything.

Given we're hopefully getting rid of the nightly toolchain very soon
anyway, let's just revert it for now rather than sinking an inordinate
amount of effort into the bump.